LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbdatabase_deleteObjectStore4-not_reused-expected.txt


PASS IDBDatabase.deleteObjectStore() - the object store is not reused
